# src/docs/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Influencerium API",
        "description": "REST API for Influencerium - Influencer Data Management Platform. Manage influencers, campaigns, data models, and analytics with secure authentication and role-based access control.",
        "version": "1.0.0",
        "contact": {
            "name": "Influencerium Support",
            "email": "support@influencerium.com",
            "url": "https://influencerium.com/help"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000/api/v1",
            "description": "Development server"
        },
        {
            "url": "https://api.influencerium.com/api/v1",
            "description": "Production server"
        }
    ],
...
    "paths": {
...
        "/influencers": {
            "get": {
                "tags": [
                    "Influencers"
                ],
                "summary": "List influencers",
                "description": "Returns a paginated list of influencers with optional filtering",
                "operationId": "listInfluencers",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "search",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Search by name, username, or email"
                    },
                    {
                        "in": "query",
                        "name": "platform",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "instagram",
                                "tiktok",
                                "youtube",
                                "twitter",
                                "facebook",
                                "linkedin",
                                "other"
                            ]
                        },
                        "description": "Filter by platform"
                    },
                    {
                        "in": "query",
                        "name": "category",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Filter by category"
                    },
                    {
                        "in": "query",
                        "name": "status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive",
                                "pending",
                                "blocked"
                            ],
                            "default": "active"
                        },
                        "description": "Filter by status"
                    },
                    {
                        "in": "query",
                        "name": "minFollowers",
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Minimum follower count"
                    },
                    {
                        "in": "query",
                        "name": "sortBy",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "name",
                                "followers",
                                "engagement_rate",
                                "created_at"
                            ],
                            "default": "created_at"
                        },
                        "description": "Field to sort by"
                    },
                    {
                        "in": "query",
                        "name": "sortOrder",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ],
                            "default": "desc"
                        },
                        "description": "Sort order"
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        },
                        "description": "Page number"
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 100
                        },
                        "description": "Items per page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of influencers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InfluencerListResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Influencers"
                ],
                "summary": "Create influencer",
                "description": "Creates a new influencer profile",
                "operationId": "createInfluencer",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateInfluencerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Influencer created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Influencer"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Influencer already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/influencers/{id}": {
            "get": {
                "tags": [
                    "Influencers"
                ],
                "summary": "Get influencer",
                "description": "Returns a single influencer by ID",
                "operationId": "getInfluencer",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "description": "Influencer ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Influencer details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Influencer"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Influencer not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Influencers"
                ],
                "summary": "Update influencer",
                "description": "Updates an existing influencer profile",
                "operationId": "updateInfluencer",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "description": "Influencer ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateInfluencerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Influencer updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Influencer"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Influencer not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Influencers"
                ],
                "summary": "Delete influencer",
                "description": "Deletes an influencer profile",
                "operationId": "deleteInfluencer",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "description": "Influencer ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Influencer deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Influencer not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/analytics/influencers": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Get influencer analytics",
                "description": "Returns influencer analytics including top performers and trends",
                "operationId": "getInfluencerAnalytics",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Influencer analytics data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InfluencerAnalytics"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT authorization header. Example: `Authorization: Bearer <token>`"
            }
        },
        "schemas": {
            "Error": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "type": "object",
                        "properties": {
                            "message": {
                                "type": "string",
                                "example": "An error occurred"
                            },
                            "code": {
                                "type": "string",
                                "example": "ERROR_CODE"
                            }
                        }
                    }
                }
            },
            "Success": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Operation successful"
                    }
                }
            },
...
            "Influencer": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "name": {
                        "type": "string",
                        "example": "Alex Johnson"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "alex@example.com"
                    },
                    "username": {
                        "type": "string",
                        "example": "alexjohnson"
                    },
                    "platform": {
                        "type": "string",
                        "enum": [
                            "instagram",
                            "tiktok",
                            "youtube",
                            "twitter",
                            "facebook",
                            "linkedin",
                            "other"
                        ],
                        "example": "instagram"
                    },
                    "profile_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://instagram.com/alexjohnson"
                    },
                    "followers": {
                        "type": "integer",
                        "example": 150000
                    },
                    "engagement_rate": {
                        "type": "number",
                        "format": "decimal",
                        "example": 4.5
                    },
                    "location": {
                        "type": "string",
                        "example": "Los Angeles, CA"
                    },
                    "bio": {
                        "type": "string",
                        "example": "Tech reviewer and lifestyle creator"
                    },
                    "category": {
                        "type": "string",
                        "example": "Technology"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "tech",
                            "lifestyle",
                            "reviews"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive",
                            "pending",
                            "blocked"
                        ],
                        "example": "active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "CreateInfluencerRequest": {
                "type": "object",
                "required": [
                    "name",
                    "platform"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 2,
                        "maxLength": 100,
                        "example": "Alex Johnson"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "username": {
                        "type": "string",
                        "minLength": 2,
                        "maxLength": 50
                    },
                    "platform": {
                        "type": "string",
                        "enum": [
                            "instagram",
                            "tiktok",
                            "youtube",
                            "twitter",
                            "facebook",
                            "linkedin",
                            "other"
                        ]
                    },
                    "profile_url": {
                        "type": "string",
                        "format": "uri"
                    },
                    "followers": {
                        "type": "integer",
                        "minimum": 0
                    },
                    "engagement_rate": {
                        "type": "number",
                        "minimum": 0,
                        "maximum": 100
                    },
                    "location": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "bio": {
                        "type": "string",
                        "maxLength": 500
                    },
                    "category": {
                        "type": "string",
                        "maxLength": 50
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "maxLength": 30
                        }
                    },
                    "notes": {
                        "type": "string",
                        "maxLength": 1000
                    }
                }
            },
...
            "InfluencerListResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "data": {
                        "type": "object",
                        "properties": {
                            "influencers": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Influencer"
                                }
                            },
                            "pagination": {
                                "$ref": "#/components/schemas/Pagination"
                            }
                        }
                    }
                }
            },
...
            "Pagination": {
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer",
                        "example": 1
                    },
                    "limit": {
                        "type": "integer",
                        "example": 20
                    },
                    "total": {
                        "type": "integer",
                        "example": 100
                    },
                    "pages": {
                        "type": "integer",
                        "example": 5
                    }
                }
            }
        }
    }
}